moved flatEdge class to own module
